+++ definitions/zemismart/registry.cpp
@@ -53,12 +53,7 @@ extern const PreparedDefinition kDef_ZM_RM02;
 
 extern const PreparedDefinition kDef_ZMS_208US_2;
 extern const PreparedDefinition kDef_ZMS_208US_3;
-extern const PreparedDefinition kDef_ZMP1;
 extern const PreparedDefinition kDef_ZMS_206EU_2;
-extern const PreparedDefinition kDef_ZM25RX_08_30;
-extern const PreparedDefinition kDef_ZM16B;
-extern const PreparedDefinition kDef_ZN_USC1U_HT;
-extern const PreparedDefinition kDef_ZM25R1;
 extern const PreparedDefinition kDef_ZMS_206US_4;
 extern const PreparedDefinition kDef_ZMS_206EU_3;
 extern const PreparedDefinition kDef_ZMS_206US_1;
@@ -110,12 +105,7 @@ const PreparedDefinition* const kZemismartRegistry[] = {
     &kDef_ZM_RM02,
     &kDef_ZMS_208US_2,
     &kDef_ZMS_208US_3,
-    &kDef_ZMP1,
     &kDef_ZMS_206EU_2,
-    &kDef_ZM25RX_08_30,
-    &kDef_ZM16B,
-    &kDef_ZN_USC1U_HT,
-    &kDef_ZM25R1,
     &kDef_ZMS_206US_4,
     &kDef_ZMS_206EU_3,
     &kDef_ZMS_206US_1,